react updates

+++ lipnonet/rekreace/api/pdo_delete_pocasi.php
@@ -1,6 +1,6 @@
 <?php 
   // Headers
-  header('Access-Control-Allow-Origin: *');
+  include_once '../config/cor.php';
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: PUT');
   header('Access-Control-Allow-Headers: Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods, Authorization, X-Requested-With');
@@ -28,22 +28,24 @@
   $webTokenCalculated = "".md5($date);
   $webToken = "".($data->webToken);
 
+  // if authentication ok
   if($webToken == $webTokenCalculated) {
-    // if update OK
-    if($post->delete()) {
-      echo json_encode( array('result' => 'pocasi_deleted') );
-      // update last update date
-      $fp = FOpen("../formular_counter.dat", "w+");
-      $LastChange=StrFTime("%d.%m.%Y");
-      FPutS($fp, $LastChange);
-      FClose($fp);
-    // if update failed
-    } else {
-      echo json_encode( array('result' => 'pocasi_not_deleted')  );
-    }
+      // if db connect ok
+      if($db){
+          // if db operation ok
+          if($post->delete()) {
+              echo json_encode( array('result' => 'pocasi_deleted') );
+              // create new graphs
+              $img_path = "../";
+              include "../graph_all.php";
+          // if update failed
+          } else {
+              echo json_encode( array('result' => 'pocasi_not_deleted')  );
+            }
+      } else {
+          echo json_encode( array('result' => 'pocasi_not_deleted - db error')  );
+        }
     // if login failed
   } else {
-    echo json_encode(
-      array('result' => 'loginFailed')
-    );
-  }
+      echo json_encode( array('result' => 'loginFailed') );
+    }
